Adobe-Hackathon-main/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp>=3.11.11
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.2
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==45.0.6
csvw==3.5.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dateparser==1.1.8
decorator==5.2.1
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpeg==1.4
filelock==3.19.1
Flask==3.1.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
g2pkk==0.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.7
httplib2==0.22.0
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
isodate==0.7.2
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.3.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
melo==1.1.0
modelscope==1.18.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy>=1.23.2,<2.1
oauthlib==3.3.1
openai-whisper==20240930
packaging==25.0
pandas>=1.5.3
pdfminer.six==20250506
pdfplumber==0.11.7
phonemizer==3.3.0
pillow>=10.4,<12
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pynndescent==0.5.13
pyparsing==3.2.3
pypdfium2==4.30.0
pypinyin==0.55.0
pysbd==0.3.4
pytesseract==0.3.13
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.11.4
segments==2.3.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.2
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.55.2
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.3.8
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
crawl4ai==0.6.0
playwright>=1.49.0